new dashboard is added
there are 5 main charts and 3 filters.
filters:
1. top 100 customers: customers who are predicted as the highest 100 amount during the time period.
2. worst 100 customers: customers who are predicted as the lowest 100 amount during the time period.
3. aggregation decision: Sum of line chart of data OR Average of line chart of data.
Graphs:
Top 100 customers: If it is not selected from filter aggregation of top 100 customers.
Worst 100 Customers: If it is not selected from the filter aggregation of the worst 100 customers
Line Chart Of CLV time period of aggregation.
When we hover to any point on the line chart it triggers the next graph which is a timeline of newcomers starting from the hovered date. Also Another timeline chart for churn customers ( customers who have no predicted value last time period and the prediction period. of time