purchase amount model data manipulation
data model must be selected according to order_count which is an input from platform. Each customer of last .. (order_count) th of order must be selected as feature to our model. Last order of each customer must be selected as Dependent variable of the Puschase Amount Model. In this case some users of orders will not be fully filled as many as order count. This will handle the problems when a customer has an order count which has less than order_count.