checking model existsing
when the model has been initialized even for training it checks if previously trained models exist in the export_path or not. It checks according to the time period. The day difference (or hour)  between the Last trained models of the trained date and the current date is bigger than the time_period, the model will be created for retraining. e.g. previous trained date 2020-12-10; time period: Month; Triggered in 2021-01-11 (day difference 31 days); the model will be retrained. Once the newly generated model has been trained, it is imported to export_path with the current date and the time period. when functions check the previous model, it checks model type (purchase amount model or next purchase model) and time period of the model. the filename of the model is assigned by model_path function. (model_name + current_Date + timePeriod).