history for purchase amount model is added - optional.
It is only needed for the epoch parameter of tuning.  It checks the last 2 epochs of loss value difference. If the difference is less than accept_threshold_for_loss_diff which is 0.05, then the epoch is assigned as the optimum one.
Who train and test data sets are normalized that is why it is expected to see the lower improvements on each epoch.