bug fixed; model save load process has issue after loading for exported model files (.json and h5 files)
I was only trying to load the current day of models. Models are saving with for the format;
   - trained_next_purchase_model_20210508_week.json
Just like the example above, if you were running for 2021-05-09 with time_period = 'week', you are able to reuse the model file above. (Fixed). Just for reminder, if you are running for 2021-05-17, the time difference will be;
    2021-05-08 (model exported date) - 2021-05-17 (prediction execution date)> 7 days (time period. For time_period = month, it would be 30 days)